requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.4.26
click==8.2.0
dj-database-url==3.0.0
Django==5.2.1
django-extensions==4.1
django-sendgrid-v5==1.3.0
ecdsa==0.19.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
greenlet==3.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==25.0
psycopg2-binary==2.9.10
PyMySQL==1.1.1
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
sendgrid==6.12.4
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.13.2
Werkzeug==3.1.3
